# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BetterPlay API",
        "description": "Aquí podrás ver todas las funciones de la API detalladamente",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pools/create": {
            "put": {
                "tags": [
                    "pools"
                ],
                "summary": "Crea una nueva quiniela",
                "description": "Recibe el nombre, la fecha final y los partidos de una quiniela y crea una nueva quiniela",
                "operationId": "e06c3fce3a3a159d2bc8da7fbe79aa80",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "matches": {
                                        "type": "string"
                                    },
                                    "finalDate": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiniela guardada correctamente"
                    }
                }
            }
        },
...
        "/api/pools/participateInPool": {
            "post": {
                "tags": [
                    "pools"
                ],
                "summary": "Participación en una quiniela",
                "description": "Recibe la id de la quiniela en la que participa, las monedas y los resultados para esta quiniela y crea una participación y recibe una notificación el usuario en cuestión",
                "operationId": "ca10f7fabd28b9960840fd9fc607e953",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "poolId": {
                                        "type": "integer"
                                    },
                                    "coins": {
                                        "type": "integer"
                                    },
                                    "poolResults": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participación creada"
                    }
                }
            }
        },
        "/api/pools/finishPool": {
            "post": {
                "tags": [
                    "pools"
                ],
                "summary": "Finalizar una quiniela",
                "description": "Recibe la id de la quiniela a finalizar y los resultados, actualiza todas las tablas y por último, envía notificación a todos los usuario que hayan participado y les suma o no las monedas según sus aciertos.",
                "operationId": "e09a00a07796edfde1edcc08f73cae6b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "poolId": {
                                        "type": "integer"
                                    },
                                    "poolResults": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participación creada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}